Replace words with bytes for sending data

+++ Firmware/PY32/src/cube.h
@@ -34,7 +34,7 @@ typedef enum
     CUBE_LEFT = 0x08
 } cube_side_t;
 
-typedef void (*cube_data_callback_t)(cube_side_t cube_side, uint32_t *data, uint32_t length);
+typedef void (*cube_data_callback_t)(cube_side_t cube_side, uint8_t *data, uint32_t length);
 typedef void (*cube_connected_callback_t)(cube_side_t cube_side);
 typedef void (*cube_disconnected_callback_t)(cube_side_t cube_side);
 typedef void (*cube_error_callback_t)(cube_side_t cube_side, cube_status_t error_code);
@@ -43,8 +43,8 @@ void cube_init();
 void cube_loop();
 uint8_t cube_is_connected(cube_side_t cube_side);
 cube_status_t cube_init_data_transfer(cube_side_t cube_side);
-void cube_send_data(cube_side_t cube_side, uint32_t *data, uint32_t length);
-cube_status_t cube_receive_data(cube_side_t cube_side, uint32_t *data, uint32_t max_length, uint32_t *length_received);
+void cube_send_data(cube_side_t cube_side, uint8_t *data, uint32_t length);
+cube_status_t cube_receive_data(cube_side_t cube_side, uint8_t *data, uint32_t max_length, uint32_t *length_received);
 
 void cube_set_data_callback(cube_data_callback_t callback);
 void cube_set_connected_callback(cube_connected_callback_t callback);
